refactor: move node package to resources package
